--- src/app/components/App.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import JobsList from './JobsList/JobsList';
-
-function App() {
-  return (
-    <div>
-      <h2>Hello JobSeekers</h2>
-      <div>
-        <JobsList />
-      </div>
-    </div>
-  );
-}
-
-export default App;
