+++ src/app/components/JobsList/JobFocusItem.jsx
