+++ src/app/components/JobsList/JobsList.jsx
@@ -4,46 +4,38 @@ import Divider from '@mui/material/Divider';
 import Paper from '@mui/material/Paper';
 import Stack from '@mui/material/Stack';
 import { styled } from '@mui/material/styles';
-import Container from '@mui/material/Container';
-import Box from '@mui/material/Box';
+import Card from '@mui/material/Card';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import Typography from '@mui/material/Typography';
 import SearchJobs from "./SearchJobs.jsx";
-import JobsList from "./JobsList.jsx";
+import JobItem from "./JobItem.jsx";
 import JobFocusItem from "./JobFocusItem.jsx";
+import Container from '@mui/material/Container';
 
 
 
 
+const JobsList = (props) => {
 
-const jobsData = {"jobs":[{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:59:49 GMT","url":"http://jobviewtrack.com/en-us/job-4e1d417e480a0601555464061e2e1e1445651a0f001d4f4e4f2f6901192157070026060607090045066a3a164442525f0d2101094e1a45061a633d13044612482c1d4d42534848166d3d421a490c1a413d13044612695e45121e/245ab288b233b8b0497e775e011ab554.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"DevOps Engineer (Senior/Staff) - San Francisco, CA","description":"DevOps <b>Engineer</b> (Senior and Staff level)   San Francisco CA (remote during Covid)   A Day in the Life of Our DevOps... <b>Engineer</b>   We own a 24/7 Global operation. Our team works in shifts and we handoff to the daytime shift in each timezone every...","company":"NextDeavor","salary":""},{"locations":"San Jose, CA","site":"","date":"Sat, 11 Dec 2021 08:59:36 GMT","url":"http://jobviewtrack.com/en-us/job-481841604c0a1a08461754161a08000045651a0f001d4f4e4f2f680a0807491145114828272e67651a0f001d4f4e4f2f6005011b411543171d1307090221435e5042/825317fce2d560bd168ef77c111f447b.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Manufacturing Engineer III","description":" conclusions.   Ability to operate a personal computer including using a Windows based operating system and related <b>software</b>....   Advanced PC skills, including training and knowledge of Jabil’s <b>software</b> packages.   Ability to write simple correspondence...","company":"Jabil","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:59:32 GMT","url":"http://jobviewtrack.com/en-us/job-4812417c58050307530d00221b121b15044e170d4936444c544348011d6c7601410f011517472453071d1b124448582f7c110e024e0059432d0f090e0b45111a6b325958485f4c0a0c0b07314e04010f0b02172231060e1a444e585f2c53595612/9f29e3a3b47e97e9c312fc50ff13738e.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Quality Assurance Engineer, Podcasts","description":"/integration <b>engineer</b>)  · Knowledge of professional <b>software</b> engineering practices & best practices for the full <b>software</b>..., and customer engagement. We are looking for new team members across a variety of job functions including <b>software</b> engineering...","company":"Amazon","salary":""},{"locations":"San Jose, CA","site":"","date":"Sat, 11 Dec 2021 08:55:07 GMT","url":"http://jobviewtrack.com/en-us/job-1d1a417e42021b19460645432d0f090e0b45111a6b205e4a5b4b0d2101094e1a45061a633d13044612483a1c4c5f4a4c5f016d2b4913490d0d041c65364f121c1e12584e3c1a1b5d5e/899a0682f878587497f63f0be076bf6f.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Staff Software Engineer, IAM","description":".   The Identity & Access Management <b>Software</b> <b>Engineer</b> is a member of the Identity Access & Management team, reporting...   Required skills:   10+ years <b>software</b> development experience in large-scale commercial networking oriented applications...","company":"Zscaler","salary":""},{"locations":"Fremont, CA","site":"","date":"Sat, 11 Dec 2021 08:54:12 GMT","url":"http://jobviewtrack.com/en-us/job-484d416a480b1b0b441c4e0a0b00024735521b020c105e0b78434a0d010b420622331a0e04020654542d0714434558485f66280b48004500000f0704044c542d0714434558485f662a00401d4e060d136c20004f000d0a1b44425e4c4165585f174d/0971f94dd03667ddc3c4f4c01b1cc6e7.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Geotechnical Project Engineer","description":"-driven and self-motivated Project <b>Engineer</b> in our Fremont office.   Education and Qualifications:   B.S. Degree in Civil... Professional <b>Engineer</b> required.  Licensed Geotechnical <b>Engineer</b> preferred.    Required Experience:   5+ years of experience in...","company":"EARTH SYSTEMS PACIFIC","salary":""},{"salary_min":"25","locations":"Concord, CA","salary_type":"H","date":"Sat, 11 Dec 2021 08:53:30 GMT","description":" and maintaining network <b>software</b> and hardware  Building and deploying file servers and cloud computing solutions  Configuring...+  MCSA (Microsoft Certified Systems Administrator/Solutions Associate) or MCSE (Microsoft Certified Systems <b>Engineer</b>) MCITP...","salary_currency_code":"USD","salary":"$25 per hour","site":"","url":"http://jobviewtrack.com/en-us/job-1f1c416148120a02073d69433c040d0f0b491701081d2862690d79010c06491d430a090f6c2e31003d3c6b274f4855434407060f49766c061e0402665319415a/35e2f5aa70e1b762d8ace0951ab74b95.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"IT Technician Level II","salary_max":"25","company":"K2 Staffing LLC"},{"locations":"Santa Clara, CA","site":"","date":"Sat, 11 Dec 2021 08:53:16 GMT","url":"http://jobviewtrack.com/en-us/job-121e417e480a06015554730c0e1519061745542d0714434558485f663d0b4a1b5406483201011157151a0c536f455a4443010a1c2527450d010e1c472145022719000a6e534a440a0a0b5576730c0e1519061745542d0714434558485f662b0b513b5010482400000c4e110d1b721d1c0c18/587432f9f6c041379b3a6de7398f2c2d.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Software / DevOps Engineer (REMOTE ELIGIBLE)","description":" operating systems, and End to End application development methodologies.   The Senior <b>Software</b> / DevOps <b>Engineer</b> works closely...% of the Fortune 500, and are the market leader in <b>Software</b>-defined Storage, Data Protection, and Archiving.   The Data...","company":"Veritas","salary":""},{"locations":"Union City, CA","site":"","date":"Sat, 11 Dec 2021 08:53:00 GMT","url":"http://jobviewtrack.com/en-us/job-1b1f4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c53575915/6bf5b7af4dec70fd7ad29d14bb1ca299.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"Sunnyvale, CA","site":"","date":"Sat, 11 Dec 2021 08:51:11 GMT","url":"http://jobviewtrack.com/en-us/job-48484169481403015e19450d1c413a0206481a010a1a4b453f79480707004e174902066327090352151b1d015f4849585f016d2a42044c0c110c0b09112143505842/dc3f335b5234a1ebb4f10adedc005e0c.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Infrastructure Deployment Technician","description":" and troubleshooting of infrastructure hardware, <b>software</b> and connectivity, all combined with a proven project delivery methodology... as a network <b>engineer</b> and deployments for large-scale networks in a corporate environment with hands on experience with physical...","company":"Amazon","salary":""},{"locations":"Pleasanton, CA","site":"","date":"Sat, 11 Dec 2021 08:48:33 GMT","url":"http://jobviewtrack.com/en-us/job-1f4e417e42021b19460645432d0f090e0b45111a6b3e434f1d6148120a0225314e04010f0b02172227070f075d4a4f482f280a18421821545d595b/9757b09986cf13c45dbc757ac3cc4dc6.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Mid-Level Software Engineer","description":"Mid-Level <b>Software</b> <b>Engineer</b>   Company: The Boeing Company   Job ID: 00000287860   Date Posted: 2021-12-09-08:00... Boeing Defense, Space & Security (BDS) business organization.   As a <b>Software</b> <b>Engineer</b> in Boeing’s Secure Computing Solutions...","company":"Boeing","salary":""},{"salary_min":"60","locations":"San Jose, CA","salary_type":"H","date":"Sat, 11 Dec 2021 08:48:25 GMT","description":"job summary:  Cloud Security <b>Engineer</b>   WHO WE ARE   Client's Customer & Partner Experience Engineering (CPXE... team.   WHAT YOU'LL DO   The CPXE Security team is seeking an <b>engineer</b> to collaborate with others in the design and build...","salary_currency_code":"USD","salary":"$60 per hour","site":"","url":"http://jobviewtrack.com/en-us/job-1e13416e410b1a0a072745001d1307131c0031060e1a444e585f2f27030152100026060607090045066a3a16495e4f44591d4f2b4913490d0d041c65264c1b1d0d53794e5e585f0d1b17252745001d1307131c21435e5042/53394e9b6541563f8e732db22bbda07c.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Cloud Security Engineer","salary_max":"60","company":"Randstad"},{"locations":"Pleasanton, CA","site":"","date":"Sat, 11 Dec 2021 08:47:55 GMT","url":"http://jobviewtrack.com/en-us/job-194a416479442a00401d4e060d136c2e31003d3c6b36444c544348011d6f10411856/d3a7453e6a8871a3c575489ebcf0c7b3.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"IT Engineer","description":"Veeva [NYSE: VEEV] is the leader in cloud-based <b>software</b> for the global life sciences industry. Committed to innovation... biotechs. Veeva's <b>software</b> helps our customers bring medicines and therapies to patients faster.   We are the first public...","company":"Veeva Systems","salary":""},{"locations":"Sunnyvale, CA","site":"","date":"Sat, 11 Dec 2021 08:45:58 GMT","url":"http://jobviewtrack.com/en-us/job-1b1e417e480a06015554730c0e1519061745542d0714434558485f44250f511522300d0f0708170027070f075d4a4f480d2101094e1a45061a633d08035403091b160a6e534a440a0a0b55546a021e006c34004e1d071b536f455a4443010a1c073e4115096324061341543b06155e5c5c5f48443c0b491d4f116956565654/e014f595b683d0a52524a7c7f7000978.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Software Engineer, Java Platform","description":" <b>software</b> development problems in principled and reusable ways.    Basic Qualifications   8+ Years of proven understanding... building reusable tools, packages and <b>software</b> patterns for developers.  Passion for <b>software</b> quality, including testing, code...","company":"Uber","salary":""},{"locations":"San Ramon, CA","site":"","date":"Sat, 11 Dec 2021 08:44:40 GMT","url":"http://jobviewtrack.com/en-us/job-4e18417e480a06015554650d0f0800020052762d0714434558485f663c0b491d4f11695659575c/bb717c094b8aed319f699f4a587d4ab2.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Engineer","description":" <b>Engineer</b> to our growing team. This is a fantastic career opportunity for a proactive, highly motivated individual..., and tickets  Manage and respond to escalated client inquiries  Prepare and implement advanced hardware and <b>software</b> for clients...","company":"Parachute Technology Inc","salary":""},{"salary_min":"120000","locations":"San Francisco, CA","salary_type":"Y","date":"Sat, 11 Dec 2021 08:42:24 GMT","description":"If you are an ASIC SOC Verification <b>Engineer</b> with experience, please read on!   We are an innovative cloud solutions... a largely funded startup, then please apply!   What You Will Be Doing   As an ASIC Verification <b>Engineer</b>, you will work with the...","salary_currency_code":"USD","salary":"$120000 - 190000 per year","site":"","url":"http://jobviewtrack.com/en-us/job-1919416c7e2d2c4e7111520a0e080d0611491b064936444c544348011d6c6627692048370b150c461d0b08074344532f7b011d07411d43021c08010945651a0f001d4f4e4f2f6c37262d07314e04010f0b02172227272a537c4e4f444b0d0c0f531d4f0d6956585f50/03fd074ae3c2ed3fc1f37c4f6efa2361.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"ASIC SOC Verification Engineer","salary_max":"190000","company":"CyberCoders"},{"locations":"Redwood City, CA","site":"","date":"Sat, 11 Dec 2021 08:41:45 GMT","url":"http://jobviewtrack.com/en-us/job-184e417e480a060155547116090d07131c0031060e1a444e585f2f351a0f4b1d541a482400000c4e110d1b71794e534442164f2b4913490d0d041c6536451a0106010a7a484c410d1b1725314e04010f0b021721435e5b4a/70286fb78ed15b9c6c04215c1ebd3b54.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Quality Engineer","description":"Job Summary & Responsibilities   The Sr. Quality <b>Engineer</b> is responsible for Quality Assurance activities ranging... from product development through transfer to manufacturing. This is a hands-on role where the Sr. Quality <b>Engineer</b>...","company":"Nevro","salary":""},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 08:39:27 GMT","url":"http://jobviewtrack.com/en-us/job-1a12417e42021b19460645432d0f090e0b45111a6b36444c544348011d6c741b46171f001c02641747515f/7e2b2438778143e07300d93b2ca5fb4e.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer","description":" on how to apply, click “Apply to Job” online on this web page.   <b>Software</b> <b>Engineer</b> Responsibilities   Help build the next generation..., Computer <b>Software</b>, Computer Engineering, Applied Sciences, Mathematics, Physics, or related field and 24 months of experience...","company":"Meta","salary":""},{"locations":"San Francisco, CA - London","site":"","date":"Sat, 11 Dec 2021 08:39:14 GMT","url":"http://jobviewtrack.com/en-us/job-4b12412c1a52575b2545125a515557/3a754a7d7746c2bef9d69001852c04c4.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior React Native Engineer - Booking Experience","description":".   As a Senior React Native <b>Engineer</b> on the Booking Experience team, you'll be a key member of our team, proposing creative solutions... with all parts of the <b>software</b> development lifecycle, including testing and release management  Strong analytical and communication...","company":"TaskRabbit","salary":""},{"locations":"Sunnyvale, CA","site":"","date":"Sat, 11 Dec 2021 08:39:06 GMT","url":"http://jobviewtrack.com/en-us/job-4e4f417e42021b19460645432d0f090e0b45111a6b36444c544348011d4e6e3d69612d0f090e0b45111a6b20454d495a4c160a6f104c1152/f19e2454017412d045a28da2503ec6c5.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer III","description":".   As a <b>Software</b> <b>Engineer</b> III, you’ll spend your days translating requirements into solutions, writing code, collaborating with cross... us on the cutting-edge of user design and a seamless customer experience. We’re intrigued by the opportunity to <b>engineer</b> the...","company":"Walmart","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:38:19 GMT","url":"http://jobviewtrack.com/en-us/job-124a417e42021b19460645432d0f090e0b45111a6b314b48564843004f2b4913490d0d041c65204e130107164f593f7e42021b19460645612a000d0c004e10695e45121e/f9c0fdaacc5035029bc848308210ccdf.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer - Backend","description":" BUT MUST BE IN THE UNITED STATES   CyberCube Analytics is looking for a <b>Software</b> <b>Engineer</b> - Backend   Our Technology Stack:   Java... servers, caching, databases, etc  Strong foundation in object-oriented programming, algorithms, and <b>software</b> application...","company":"Cybcube","salary":""},{"locations":"Sunnyvale, CA","site":"","date":"Sat, 11 Dec 2021 08:34:44 GMT","url":"http://jobviewtrack.com/en-us/job-4f1d417e42021b19460645433c041d1345651a0f001d4f4e4f2f7e0b091a50155206482400000c4e110d1b717e4e4e590d2101094e1a45061a633d08035403091b160a7f585e59662a00401d4e060d136f505d1145/adf1415f8a06cd8f81724d41ca81df6e.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Test Engineer","description":"The <b>Software</b> Test <b>Engineer</b> develops <b>software</b> used in mass production environments to calibrate, measure, and interface.... We are looking for someone with a productive mindset who can apply their expertise to solve problems in creative, insightful ways.   <b>Software</b> Test <b>Engineer</b>...","company":"Meta","salary":""},{"salary_min":"65000","locations":"Daly City, CA","salary_type":"Y","date":"Sat, 11 Dec 2021 08:30:36 GMT","description":" Commercial Drywall Contractor seeking an experienced Project <b>Engineer</b>. This is an exciting opportunity to become a part... industry ready to learn. Commercial Drywall Project <b>Engineer</b> Experience Required. If you have an interest in the Commercial...","salary_currency_code":"USD","salary":"$65000 - 80000 per year","site":"","url":"http://jobviewtrack.com/en-us/job-1b19417d5f0b050b44000026060607090045066a2c1d4d42534848166d3e551b4a060b156c2317590309051f2b1d04151e/653b799d0a24040d0b55fd247896ffca.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Drywall Project Engineer: $65K-$80K","salary_max":"80000","company":"GPAC"},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 08:28:18 GMT","url":"http://jobviewtrack.com/en-us/job-1a12417e42021b19460645432d0f090e0b45111a6b36444c544348011d6c741b46171f001c02641747515f/385eae4c14beb8657519fc3d40d98a3e.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer","description":" on how to apply, click “Apply to Job” online on this web page.   <b>Software</b> <b>Engineer</b> Responsibilities   Help build the next generation..., Computer <b>Software</b>, Computer Engineering, Applied Sciences, Mathematics, Physics, or related field. Position also requires...","company":"Meta","salary":""},{"locations":"Santa Clara, CA","site":"","date":"Sat, 11 Dec 2021 08:27:39 GMT","url":"http://jobviewtrack.com/en-us/job-4e4a417e480a06015554730c0e1519061745542d0714434558485f663c0b491d4f1148321714114519482c1d4d42534848166d3d4812541409130b47204e130107164f591d6a5f051f064e1753613b0e08131241060d49205358494840442a00401d4e060d136c340a46001f08014f0b78434a0d010b420621545f505b/6e7a1fdac945fe6f72d9dc9ddad0e645.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Graphics System Software Engineer","description":"NVIDIA is searching for outstanding <b>software</b> engineers and architects for an exciting role in our Tegra <b>software</b>..., OpenGL ES, Vulkan, EGL drivers, and extensions   Build new <b>software</b> functions and/or expose new hardware functionalities in...","company":"Nvidia","salary":""},{"locations":"Foster City, CA","site":"","date":"Sat, 11 Dec 2021 08:26:58 GMT","url":"http://jobviewtrack.com/en-us/job-1a1f417b4c08060a4600490c06412b0902491a0d0c0128795c494c164f2b4913490d0d041c65204e130107164f593f7b4c08060a4600490c06633c0601410669584a181b0b15/93e6a9aa94f2e56269133388a3e248f7.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Radar Validation Engineer","description":" to be part of a fast-moving and highly execution-oriented team. Zoox is looking for a Radar Validation <b>Engineer</b> with a background... radar performance  Experience with hands-on radar troubleshooting for problems that intersect hardware and <b>software</b>  Cross...","company":"Zoox","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:25:16 GMT","url":"http://jobviewtrack.com/en-us/job-4e4e4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c53595612/56eeea599b4562712a1274f43e26f553.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer","description":" of our Value-Based Payment product suite to customers.   YOUR IMPACT   As a Technical Support <b>Engineer</b>, you will work with a team... with our customers by engaging via email, phone, screen share, and CRM <b>Software</b> (e.g. Zendesk).  Track and report performance on Support...","company":"Nuna","salary":""},{"locations":"Los Altos, CA","site":"","date":"Sat, 11 Dec 2021 08:25:12 GMT","url":"http://jobviewtrack.com/en-us/job-4e4a4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c535c5a16/1056f30d774c94ca557a74131878017f.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"Santa Clara, CA","site":"","date":"Sat, 11 Dec 2021 08:21:45 GMT","url":"http://jobviewtrack.com/en-us/job-1b4a417e42021b19460645432d0f090e0b45111a493a445f585f43663c01410057021a044e220b471d060c16580b107e54171b0b4a07223011121a020853542d0714434558485f4426005311520d6a2202081044543b06155e5c5c5f48442a00401d4e060d136c340a46001f08014f0b78434a0d010b420621545f505b/437e3880a4c17c42a877b995f6202596.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Intern - Cloud Software Engineer (Threat Research Systems)","description":"Company Description   Our Mission   At Palo Alto Networks® everything starts and ends with our mission:   Being the cybersecurity partner of choice, protecting our digital way of life.   We have the vision of a world where each day is safer...","company":"Palo Alto Networks","salary":""},{"locations":"Mountain View, CA","site":"","date":"Sat, 11 Dec 2021 08:20:46 GMT","url":"http://jobviewtrack.com/en-us/job-1b184179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c535b5a1e/e509646b01ea54452290076802affbe7.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"Milpitas, CA","site":"","date":"Sat, 11 Dec 2021 08:18:24 GMT","url":"http://jobviewtrack.com/en-us/job-134d417f48170a0f551748432d0f090e0b45111a6b374f5d58414214020b49000026060607090045066a21165d47585959443f0f441f41110c633c021645151a0a1b0a675c4f5e662e0a51154e000d054e23005611040603474e53592c535b5e1e/1994d524599535e49efbf6b6c5a290b2.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Hewlett Packard Labs - Research Engineer/Advanced Development","description":"Hewlett Packard Labs – Research <b>Engineer</b>/Advanced Development   This role has been designated as ‘Office’, which means... that validate their research.   About The Job   Primary areas of focus will be in: a) <b>software</b> frameworks, programming models...","company":"Hewlett Packard Enterprise","salary":""},{"locations":"Fremont, CA","site":"","date":"Sat, 11 Dec 2021 08:18:07 GMT","url":"http://jobviewtrack.com/en-us/job-4f1b4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c535e5e1e/e466183bc327541cf131c0ffaa2f6660.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"San Jose, CA","site":"","date":"Sat, 11 Dec 2021 08:17:45 GMT","url":"http://jobviewtrack.com/en-us/job-4b4a416841010c1a551d430204412a02164913064936444c544348011d6c621845001c130704044c542d0714434558485f662b0b541d470d482400000c4e110d1b716f47584e5916060d461800270d1207000b2231060e1a444e585f2c53595716/6c0f2bf26eb451ebcef2592fb5129931.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Electrical Design Engineer","description":"&D think-tank is looking for a hands-on Electrical <b>Engineer</b> to join their team! ! In this role you will design, prototype... and test the next generation of flexible hybrid electronics. This is a unique role perfect for an <b>engineer</b> who is interested in...","company":"Manpower","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:16:18 GMT","url":"http://jobviewtrack.com/en-us/job-4c1a417e42021b19460645432d0f090e0b45111a6b36444c544348011d6c741b46171f001c02641742505c/e1526c894893ee7caee9aa6cf25b04e0.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer 3 - San Francisco, CA","description":"<b>Software</b> <b>Engineer</b> 3  6+ months W2 Contract  San Francisco, CA, or Boulder, CO or Portland, OR Area Preferred (remote... currently during COVID)   The Challenge:  We are looking for a DevOps <b>Engineer</b> who loves building sophisticated and highly...","company":"NextDeavor","salary":""},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 08:14:00 GMT","url":"http://jobviewtrack.com/en-us/job-4f49416348101801551f0026060607090045066a2a16464748414c164f2b4913490d0d041c65204e130107164f593f6348101801551f22200d0d0212094106695e40131d/03d680006a8598e49bbcdf36bf141149.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Network Engineer, Cellular","description":" Reality Labs.   Network <b>Engineer</b>, Cellular Responsibilities   Planning and designing multiple test lanes, space, mechanical... for multiple test lanes, testbeds   Partner cross-functionally with device teams, hardware/<b>software</b> teams, certifications team...","company":"Meta","salary":""},{"locations":"Sunnyvale, CA","site":"","date":"Sat, 11 Dec 2021 08:13:59 GMT","url":"http://jobviewtrack.com/en-us/job-1a4e417c58050307530d00221b121b15044e170d4936444c544348011d6c7601410f01151747204e130107164f591d6464663e1b4618491711412f141655060907104f296c584c08061a5e54650d0f080002005276291a005f595c434e014f2b4913490d0d041c6652184559/1aebc5cb8b2b2cbdca159eee9cdfc7a5.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Quality Assurance Engineer II","description":", or as a programmer/analyst, quality assurance <b>engineer</b>, <b>software</b> test developer, or a related occupation.   In the alternate, employer..., or a related field and five years of experience in the job offered or as a programmer/analyst, quality assurance <b>engineer</b>, <b>software</b>...","company":"Amazon","salary":""},{"locations":"San Ramon, CA","site":"","date":"Sat, 11 Dec 2021 08:12:21 GMT","url":"http://jobviewtrack.com/en-us/job-181b416f44034f2a460041432d0f090e0b45111a6b20454d495a4c160a4e621a470a06040b156764151c08536f455a4443010a1c2536490448250f13042231060e1a444e585f2c53585e1e/344314a6594314bda9216a5ddb32ae5f.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Big Data Software Engineer WFH","description":"’t just imagine the future – you’ll create it.   As a Big Data <b>Software</b> <b>Engineer</b>, you'll be responsible for the...","company":"AT&T","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:10:47 GMT","url":"http://jobviewtrack.com/en-us/job-1f1b41604c0707074911002f0d001c090c4e13482c1d4d42534848166d234617480a06044e2b00410606001d4d296e5f0d2101094e1a45061a63230606481d060c536f455a4443010a1c25314e04010f0b0217003d2168441c1308/cb5cf30cda106e2ca4ce207ee6fed177.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Sr Machine Learning Engineer II - Maps","description":" <b>engineer</b> on this team you would help us build out the ML models that drive everything from ETA calculations to determining the... mapping professionals, data scientists, <b>software</b> engineers, and research scientists on a user-facing products with global...","company":"Uber","salary":""},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 08:09:55 GMT","url":"http://jobviewtrack.com/en-us/job-1a12417e42021b19460645432d0f090e0b45111a6b36444c544348011d6c741b46171f001c02641747515f/6831db0a0f47246ac37501f38e8de5dc.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer","description":" on how to apply, click “Apply to Job” online on this web page.   <b>Software</b> <b>Engineer</b> Responsibilities   Help build the next generation..., Computer <b>Software</b>, Computer Engineering, Applied Sciences, Mathematics, Physics, or related field and 24 months of experience...","company":"Meta","salary":""},{"locations":"San Leandro, CA","site":"","date":"Sat, 11 Dec 2021 08:06:07 GMT","url":"http://jobviewtrack.com/en-us/job-4f1b4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c53595713/6f446d4f4dd325fd54be42b01294e3bb.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"Santa Clara, CA","site":"","date":"Sat, 11 Dec 2021 08:05:43 GMT","url":"http://jobviewtrack.com/en-us/job-1e194179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c53585f12/bde1421d4e59b4b2455b3c0abdbc9ca2.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 08:05:31 GMT","url":"http://jobviewtrack.com/en-us/job-194d417d5f0b0b1b4400002e0913050211491a0f49374359584e590b1d6c6a1552080d150709020030011b16495f525f2f360a034800454325001c0c00541d060e717a59524958071b4e6a1552080d1507090222241a06175f48490d690d1d0b44004f116956585f50/649d5cf4fd65dbc1ea45522cc006a3fa.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Director, Product Marketing - Remote","description":"Harness is a rapidly growing startup that is disrupting the <b>software</b> delivery market. We are building an intelligent... <b>software</b> delivery platform that enables engineers to deliver <b>software</b> faster, with higher quality, and less effort. The Harness...","company":"Harness","salary":""},{"locations":"Newark, CA","site":"","date":"Sat, 11 Dec 2021 08:04:08 GMT","url":"http://jobviewtrack.com/en-us/job-181b4179480707004e17410f48321b17154f061c4936444c544348011d6c7311430b06080d060900271d19034559492f7e111f1e480654432d0f090e0b45111a6b274f48554344070e0207314e04010f0b02172231060e1a444e585f2c535b5816/c6785362101e1a9e63f7d69dedda321a.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Technical Support Engineer - Fremont, CA","description":"As a Technical Support <b>Engineer</b>, you will be responsible for all activities relating to Test Engineering (<b>Software</b>... support over phone and/or web to customers on hardware and <b>software</b> issues. Goal is to Drive customer satisfaction through...","company":"Express Employment Professionals","salary":""},{"locations":"Burlingame, CA","site":"","date":"Sat, 11 Dec 2021 08:04:07 GMT","url":"http://jobviewtrack.com/en-us/job-4e1b417a440a0b015007003011121a020853542d0714434558485f663c175400450e1b412b0902491a0d0c010a62742f7e1d1c1a421953432d0f090e0b45111a6b24434559425a174f2b4913490d0d041c6532491a0c0604590b6e545e100a035475165b5c54/c0e84bb9a39f9db902c166549a8fb81e.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Systems Engineer II - Windows","description":" <b>software</b> and analytic tools, domain expertise, and our unique results-oriented approach. Customers worldwide use our services... Systems <b>Engineer</b> II - Windows support the technology environments and are responsible for building, supporting...","company":"Argus","salary":""},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 08:02:17 GMT","url":"http://jobviewtrack.com/en-us/job-1a12417e42021b19460645432d0f090e0b45111a6b36444c544348011d6c741b46171f001c02641747515f/24ee9ab73ed4e65c3db875f4bb25cdd9.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer","description":" on how to apply, click “Apply to Job” online on this web page.   <b>Software</b> <b>Engineer</b> Responsibilities   Help build the next generation..., Computer <b>Software</b>, Computer Engineering, Applied Sciences, Mathematics, Physics, or related field. Position also requires...","company":"Meta","salary":""},{"locations":"Redwood City, CA","site":"","date":"Sat, 11 Dec 2021 08:01:11 GMT","url":"http://jobviewtrack.com/en-us/job-1a4d41644005080b0724520c0b041d140c4e13482c1d4d42534848166d274a15470648311c080645071b001d4d296e594c02094e621a470a06040b15677006070a16595854434a442a00401d4e060d136c2e0841130d4936444c544348011d6f1042125a/020a3097eaf7e77b3977204bdcf6dab9.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Staff Image Processing Engineer","description":"Job Description:   Employer: Auris Health, Inc.   Job Title: Staff Image Processing <b>Engineer</b>   Job Code : A011.... Develop <b>software</b> tools needed for optimization process. Optimize pipeline parameters for excellent image quality. Prepare...","company":"Johnson & Johnson","salary":""},{"locations":"San Jose, CA","site":"","date":"Sat, 11 Dec 2021 08:00:39 GMT","url":"http://jobviewtrack.com/en-us/job-1b49417e480a06015554731a1b150b0a160031060e1a444e585f2f370a004e1b52432c041d0e024e542d0714434558485f663c175400450e1b412a02164913064936444c544348011d6c740d53170d0c1d47204e130107164f593f69481706094954650d0f0800020052755f5f4a1b/01679e570b382b5e008f9bffc85fcf4c.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Staff Systems Design Engineer","description":" that improves the way we live and work. We are ONEXILINX.   Job Description:   The System <b>Software</b>, Integration & Validation group... with programmable logic, with quad core ARM processing system and programmable logic.   We are looking for a Systems Design <b>Engineer</b>...","company":"Xilinx","salary":""},{"locations":"San Mateo, CA","site":"","date":"Sat, 11 Dec 2021 07:57:19 GMT","url":"http://jobviewtrack.com/en-us/job-4c4d417d5f0d190f440d000206054e340043011a000753296e484e111d07530d003007071a10045211482c1d4d42534848166d3d4812541409130b47204e130107164f593f7e48071a1c4e0059432d0f090e0b45111a6b20454d495a4c160a4e741143161a081a1e641743585b/d8144f8743865562da9e19237cd5aaca.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Software Engineer - Security and Privacy","description":" or equivalent  7+ years of experience as a <b>software</b> <b>engineer</b> or security <b>engineer</b> that writes code  Demonstrable coding skills... and visitor management - provide seamless visibility through an integrated secure cloud-based <b>software</b> platform. Over...","company":"Verkada","salary":""},{"locations":"San Francisco, CA","site":"","date":"Sat, 11 Dec 2021 07:56:02 GMT","url":"http://jobviewtrack.com/en-us/job-4b1d417e480a06015554701107061c060800390907124d4e4f2f79010c06491d430204413e150a4706090453674a534c4a011d6c74114e0a07134e3300431c0600104b471d604c0a0e09420622331a0e0915044d5425081d4b4c585f2f370a004e1b524325000006024506695e45121e/92d994005180f3e158bff5d60f0127ab.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Technical Program Manager","description":" designs, challenges, and risks.  5+ years as a TPM/EPM or Engineering Manager or <b>Engineer</b> in a <b>software</b>, SaaS, or systems... enterprise technology spanning <b>software</b> development, data integration, FinTech, and more.   What You Will Do   You will design...","company":"Uber","salary":""},{"locations":"San Jose, CA","site":"","date":"Sat, 11 Dec 2021 07:54:53 GMT","url":"http://jobviewtrack.com/en-us/job-1213417e480a06015554730c0e1519061745542d0714434558485f442c6c64546c0a0614164731632448202328675443581c4f2d07274f051c160f15000031060e1a444e585f2f370a004e1b52433b0e08131241060d4936444c544348011d6c741b46171f001c0245651a0f001d4f4e4f0d6e6558581e45/e0fca4277acab6833e04a77d899ebd50.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Senior Staff Software Engineer (SK-ZPA) - C / Linux/ TCP/IP","description":".   Job Description   As a Senior Staff <b>Software</b> <b>Engineer</b>, you will play a crucial role in designing, creating, and enhancing multi-tenant systems...","company":"Zscaler","salary":""},{"locations":"Menlo Park, CA","site":"","date":"Sat, 11 Dec 2021 07:54:33 GMT","url":"http://jobviewtrack.com/en-us/job-4848417d5f0b0b1b4400490c06412b0902491a0d0c01286e534a440a0a0b5576701107051b0411491b06684419120b/5062dbfed4c22e48f3b15cf01a603320.html?affid=2e32adaedb8f853492180d1569e8bec5","title":"Production Engineer","description":" “Apply to Job” online on this web page.   Production <b>Engineer</b> Responsibilities   Participate in the design, implementation and ongoing..., highly-scalable performant solutions   <b>Software</b> development tools: Code editors (VIM or Emacs) and revision control systems...","company":"Meta","salary":""}],"hits":10030,"response_time":0.342247009277344,"type":"JOBS","pages":201};
 
-
-const Main = () => {
-  const [jobsState, setJobs] = useState(jobsData.jobs);
-  const [jobFocusState, setItem] = useState({})
-
-  const mainFocusFunction = (targetJobItem) => {
-    setItem(targetJobItem);
-    console.log('targetJobItem from main', targetJobItem)
-    console.log('state from main', jobFocusState)
+  const handleFocusItem = (targetJobItem) => {
+    console.log('nm, you?', targetJobItem);
+    props.mainFocusFunction(targetJobItem);
   }
-  useEffect(() => {
-    console.log('useEffect in main job state', jobFocusState);
-
-  },[]);
 
   return (
 
-      <Container sx={{ display: 'flex' }}>
-        <Box >
-          <JobsList mainFocusFunction={mainFocusFunction} listOfJobs={jobsData.jobs}/>
-        </Box>
-        <Box >
-        <JobFocusItem job={jobFocusState}  />
-        </Box>
+      <Container >
+      {props.listOfJobs.map((job) =>
+        <JobItem  key ={job.url} job={job} handleFocus={handleFocusItem}/>
+
+      )}
       </Container>
+
 
 
   )
 }
 
-export default Main;
+export default JobsList;
